Merge 0.52.0/sync into master

+++ src/sgame/components/EggComponent.cpp
@@ -47,7 +47,7 @@ bool EggComponent::CheckSpawnPoint(
 			maxDistance = VectorLength(spawnerMins) + VectorLength(clientMaxs);
 
 			// The fraction of the angle seperating best (0°) and worst case (45°).
-			frac = acos(-spawnerNormal[2]) / M_PI_4;
+			frac = acosf(-spawnerNormal[2]) / M_PI_4;
 		} else {
 			// Best case: Client spawning right above egg.
 			minDistance = spawnerMaxs[2] - clientMins[2];
@@ -56,7 +56,7 @@ bool EggComponent::CheckSpawnPoint(
 			maxDistance = VectorLength(spawnerMaxs) + VectorLength(clientMins);
 
 			// The fraction of the angle seperating best (0°) and worst case (45°).
-			frac = acos(spawnerNormal[2]) / M_PI_4;
+			frac = acosf(spawnerNormal[2]) / M_PI_4;
 		}
 
 		// The linear interpolation of min & max distance should be an upper boundary for the
